Better Organize files - Changed Routing in App.js - added dedicated manageOrdersPage - Removed unncessary files and folders
