+++ src/main/react-frontend/src/pages/ManageOrdersPage.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+export default function ManageOrdersPage() {
+	return (
+		<>
+            <h1>Manage Orders</h1>
+        </>
+	);
+}
